chore: simplify sidebar, cleanup
