+++ packages/morph/components/rails.tsx
@@ -1,0 +1,337 @@
+"use client"
+
+import * as React from "react"
+import { memo, useState, useCallback, useMemo } from "react"
+import {
+  LayoutIcon,
+  PlusIcon,
+  ChevronDownIcon,
+  ChevronRightIcon,
+  PinLeftIcon,
+  PinRightIcon,
+  CrumpledPaperIcon,
+} from "@radix-ui/react-icons"
+import { setFile } from "@/components/markdown-inline"
+import { cn } from "@/lib/utils"
+import {
+  Sidebar,
+  SidebarMenu,
+  SidebarGroupContent,
+  SidebarGroup,
+  SidebarContent,
+  useSidebar,
+  SidebarSeparator,
+} from "@/components/ui/sidebar"
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip"
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
+import { SidebarMenuButton, SidebarMenuItem, SidebarMenuSub } from "@/components/ui/sidebar"
+import { useRouter } from "next/navigation"
+import { FileSystemTreeNode, Vault } from "@/db"
+import { EditorView } from "@codemirror/view"
+import { useToast } from "@/hooks/use-toast"
+import { VaultButton } from "@/components/ui/button"
+
+interface FileTreeNodeProps {
+  node: FileSystemTreeNode
+  onFileSelect?: (node: FileSystemTreeNode) => void
+}
+
+// TODO: reducer and context for states
+// https://react.dev/learn/scaling-up-with-reducer-and-context
+export const FileTreeNode = memo(function FileTreeNode({ node, onFileSelect }: FileTreeNodeProps) {
+  const [isOpen, setIsOpen] = useState(node.isOpen ?? false)
+
+  const toggleOpen = useCallback(() => {
+    setIsOpen((prev) => !prev)
+  }, [])
+
+  const handleFileClick = useCallback(() => {
+    onFileSelect?.(node)
+  }, [onFileSelect, node])
+
+  // Filter out non-markdown files
+  if (node.kind === "file" && node.extension !== "md") {
+    return null
+  }
+
+  const MemoizedSidebarFolderItem = useMemo(
+    () => (
+      <SidebarMenuItem>
+        <CollapsibleTrigger asChild className="cursor-pointer">
+          <SidebarMenuButton onClick={toggleOpen} className="w-full text-xs">
+            {isOpen ? (
+              <ChevronDownIcon className="transition-transform w-3.5 h-3.5 mr-1 shrink-0" />
+            ) : (
+              <ChevronRightIcon className="transition-transform w-3.5 h-3.5 mr-1 shrink-0" />
+            )}
+            <span className="truncate">{node.name}</span>
+          </SidebarMenuButton>
+        </CollapsibleTrigger>
+        <CollapsibleContent>
+          <SidebarMenuSub className="mr-0 pl-3">
+            {node.children &&
+              node.children.map((child) => (
+                <FileTreeNode
+                  key={`${child.name}.${child.extension}`}
+                  node={child}
+                  onFileSelect={onFileSelect}
+                />
+              ))}
+          </SidebarMenuSub>
+        </CollapsibleContent>
+      </SidebarMenuItem>
+    ),
+    [isOpen, toggleOpen, node, onFileSelect],
+  )
+
+  const MemoizedSidebarFileItem = useMemo(
+    () => (
+      <SidebarMenuButton
+        className="data-[active=true]:bg-transparent hover:bg-accent/50 transition-colors flex items-center cursor-pointer w-full text-xs py-1"
+        onClick={handleFileClick}
+      >
+        <span className="truncate">{node.name}</span>
+      </SidebarMenuButton>
+    ),
+    [handleFileClick, node],
+  )
+
+  if (node.kind === "file") return MemoizedSidebarFileItem
+
+  return (
+    <Collapsible open={isOpen} onOpenChange={setIsOpen} className="group/collapsible" asChild>
+      {MemoizedSidebarFolderItem}
+    </Collapsible>
+  )
+})
+
+interface RailsProps extends React.ComponentProps<typeof Sidebar> {
+  vault: Vault
+  editorViewRef: React.RefObject<EditorView | null>
+  onFileSelect?: (handle: FileSystemFileHandle) => void
+  onNewFile?: () => void
+  onContentUpdate?: (content: string) => void
+}
+
+export default memo(function Rails({
+  className,
+  vault,
+  editorViewRef,
+  onFileSelect,
+  onNewFile,
+  onContentUpdate,
+  ...props
+}: RailsProps) {
+  const { toast } = useToast()
+  const { toggleSidebar, state } = useSidebar()
+  const isExpanded = state === "expanded"
+  const router = useRouter()
+
+  const onManageVault = useCallback(() => {
+    router.push("/vaults")
+  }, [router])
+
+  const onNewFileClick = useCallback(() => {
+    if (editorViewRef.current) {
+      editorViewRef.current.dispatch({
+        changes: {
+          from: 0,
+          to: editorViewRef.current.state.doc.length,
+          insert: "",
+        },
+        effects: setFile.of("Untitled"),
+      })
+    }
+    onNewFile?.()
+  }, [editorViewRef, onNewFile])
+
+  const handleIconClick = useCallback(
+    (id: string) => {
+      switch (id) {
+        case "new":
+          onNewFileClick()
+          break
+        case "home":
+          onManageVault()
+          break
+        default:
+          break
+      }
+    },
+    [onNewFileClick, onManageVault],
+  )
+
+  const handleFileSelect = useCallback(
+    async (node: FileSystemTreeNode) => {
+      if (!vault || node.kind !== "file" || !editorViewRef.current) return
+      if (node.extension !== "md") {
+        toast({ title: "File picker", description: "Can only open markdown files" })
+        return
+      }
+
+      try {
+        const file = await node.handle.getFile()
+        const content = await file.text()
+
+        if (onFileSelect) onFileSelect(node.handle as FileSystemFileHandle)
+
+        editorViewRef.current.dispatch({
+          changes: {
+            from: 0,
+            to: editorViewRef.current.state.doc.length,
+            insert: content,
+          },
+          effects: setFile.of(file.name),
+        })
+
+        if (onContentUpdate) {
+          onContentUpdate(content)
+        }
+      } catch (error) {
+        console.error("Error reading file:", error)
+      }
+    },
+    [vault, editorViewRef, onFileSelect, onContentUpdate, toast],
+  )
+
+  const width = isExpanded ? "16rem" : "3.05rem"
+
+  // TODO: Add settings panel back
+  return (
+    <TooltipProvider delayDuration={300}>
+      <div
+        className={cn("fixed z-sidebar lg:sticky h-full", className)}
+        style={{ width }}
+        {...props}
+      >
+        <nav
+          className={cn(
+            "h-screen flex flex-col items-center py-4 gap-3 fixed top-0 left-0 z-20 transition duration-100 border-border border-r",
+            "shadow-xl !bg-bg-300 backdrop-blur-sm",
+          )}
+          style={{ width }}
+        >
+          <div className="flex w-full items-center gap-px px-2">
+            <VaultButton
+              onClick={toggleSidebar}
+              color="none"
+              className={cn(
+                "transition-transform relative group",
+                "text-gray-700 hover:bg-gray-100/60",
+              )}
+              title={isExpanded ? "Collapse sidebar" : "Expand sidebar"}
+            >
+              {isExpanded ? (
+                <>
+                  <PinLeftIcon className="h-4 w-4 absolute opacity-0 group-hover:opacity-100 transition-opacity" />
+                  <LayoutIcon className="h-4 w-4 group-hover:opacity-0 transition-opacity" />
+                </>
+              ) : (
+                <>
+                  <PinRightIcon className="h-4 w-4 absolute opacity-0 group-hover:opacity-100 transition-opacity" />
+                  <LayoutIcon className="h-4 w-4 group-hover:opacity-0 transition-opacity" />
+                </>
+              )}
+            </VaultButton>
+          </div>
+          <div className="flex flex-col align-center h-full w-full overflow-hidden">
+            <div className="flex flex-col px-2 pt-1 gap-2 mb-6">
+              <div className="relative group">
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <VaultButton
+                      onClick={() => handleIconClick("new")}
+                      color="none"
+                      className={cn(
+                        "group w-full text-sm rounded-lg transition ease-in-out active:!scale-100 whitespace-nowrap",
+                        "flex !justify-start !min-w-0 px-4 py-2 h-9",
+                        "hover:bg-orange-100/20 active:!bg-orange-200/30",
+                      )}
+                      title="New file"
+                    >
+                      <div className="-mx-2 flex flex-row items-center gap-3 text-orange-600">
+                        <div className="size-4 flex items-center justify-center">
+                          <div className="p-1.5 group-active:!scale-[0.98] group-active:!shadow-none group-hover:-rotate-2 group-active:rotate-3 rounded-full transition-all ease-in-out bg-orange-100 group-hover:shadow-md">
+                            <PlusIcon className="h-3 w-3 shrink-0 transition" />
+                          </div>
+                        </div>
+                        <span
+                          className={cn("truncate text-sm whitespace-nowrap mask-image-text")}
+                          style={{
+                            opacity: isExpanded ? 1 : 0,
+                            width: isExpanded ? "auto" : 0,
+                            overflow: "hidden",
+                          }}
+                        >
+                          New file
+                        </span>
+                      </div>
+                    </VaultButton>
+                  </TooltipTrigger>
+                  <TooltipContent side="right" hidden={isExpanded}>
+                    New file
+                  </TooltipContent>
+                </Tooltip>
+              </div>
+              <div className="relative group">
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <VaultButton
+                      onClick={() => handleIconClick("home")}
+                      color="none"
+                      className={cn(
+                        "group w-full text-sm rounded-lg transition duration-300 ease-[cubic-bezier(0.165,0.85,0.45,1)]",
+                        "hover:bg-gray-100/60 overflow-hidden !min-w-0 active:bg-gray-200/70 active:scale-[0.99] px-4 py-2 h-9",
+                        "text-gray-700 hover:text-gray-900",
+                      )}
+                      title="Home"
+                    >
+                      <div className="-translate-x-2 w-full flex flex-row items-center justify-start gap-3">
+                        <div className="size-4 flex items-center justify-center group-hover:!text-gray-900 text-gray-700">
+                          <CrumpledPaperIcon className="h-4 w-4 shrink-0 group-hover:-translate-y-[0.5px] transition group-active:translate-y-0" />
+                        </div>
+                        <span
+                          className="truncate text-sm whitespace-nowrap mask-image-text"
+                          style={{
+                            opacity: isExpanded ? 1 : 0,
+                            width: isExpanded ? "auto" : 0,
+                            overflow: "hidden",
+                          }}
+                        >
+                          Home
+                        </span>
+                      </div>
+                    </VaultButton>
+                  </TooltipTrigger>
+                  <TooltipContent side="right" hidden={isExpanded}>
+                    Home
+                  </TooltipContent>
+                </Tooltip>
+              </div>
+            </div>
+
+            {state === "expanded" && (
+              <div className="transition-all duration-200 flex flex-grow flex-col overflow-y-auto overflow-x-hidden relative px-2 mb-2">
+                <h3 className="text-text-300 flex items-center gap-1.5 text-xs text-muted-foreground select-none pb-2 pl-2 sticky top-0 z-10 bg-gradient-to-b from-bg-200 from-50% to-bg-200/40">
+                  Files
+                </h3>
+                <SidebarContent className="w-full p-0">
+                  <SidebarGroup className="p-0">
+                    <SidebarGroupContent>
+                      <SidebarMenu>
+                        {vault &&
+                          vault.tree!.children!.map((node, idx) => (
+                            <FileTreeNode key={idx} node={node} onFileSelect={handleFileSelect} />
+                          ))}
+                      </SidebarMenu>
+                    </SidebarGroupContent>
+                  </SidebarGroup>
+                </SidebarContent>
+              </div>
+            )}
+          </div>
+        </nav>
+      </div>
+    </TooltipProvider>
+  )
+})
