--- packages/morph/components/sidebar/index.ts
@@ -1,1 +1,0 @@
-export { SidebarRails } from "./sidebar-rails"
