+++ packages/morph/components/ui/sidebar.tsx
@@ -357,7 +357,7 @@ function SidebarFooter({ className, ...props }: React.ComponentProps<"div">) {
     <div
       data-slot="sidebar-footer"
       data-sidebar="footer"
-      className={cn("flex flex-col gap-2 p-2", className)}
+      className={cn("lex flex-col gap-2 p-2", className)}
       {...props}
     />
   )
