+++ packages/morph/hooks/use-persisted-settings.ts
@@ -13,7 +13,6 @@ export interface Settings {
     format: "biblatex" | "csl-json"
     databasePath?: string
   }
-  showFooter: boolean
 }
 
 export default function usePersistedSettings() {
